requirements.txt
flask
selenium
git+https://github.com/Yelp/Testify.git@c9fd9e3808
cssselect
coverage
pymiproxy
proxpy